requirements.txt
streamlit
sentence-transformers
faiss-cpu
lxml
grobid-client==0.8.0
numpy